pyproject.toml
[tool.poetry]
name = "workshop"
version = "0.1.0"
description = "Workshop Wagtail site"
authors = ["Torchbox <hello@torchbox.com>"]

[tool.poetry.dependencies]
python = "~3.12"
dj-database-url = "~0.5"
django = "~4.2"
django-basic-auth-ip-whitelist = "~0.5"
django-birdbath = "^1.0.0"
django-crispy-forms = "^2.1"
django-csp = "~3.6"
django-defender = "0.9.5"
django-extensions = "~3.2"
django-pattern-library = "~1.1"
django-redis = "~5.2"
django-storages = { version = "~1.12", extras = ["boto3"] }
djhtml = "~3.0.6"
gunicorn = "~20.1"
psycopg2 = "~2.9"
scout-apm = "~2.25"
sentry-sdk = "~1.5"
tbxforms = "3.0.0"
wagtail = "~5.2"
wagtail-accessibility = "~0.2"
wagtail-django-recaptcha = { git = "https://github.com/torchbox-forks/wagtail-django-recaptcha", tag = "2.1.1+tbx" }
wagtail-factories = "^4.1.0"
wagtail-storages = "~1.0.0"
whitenoise = "~6.2"
boto3 = "^1.34.55"

[tool.poetry.group.dev.dependencies]
Werkzeug = "^2.0.3"
black = "24.2.0"
coverage = "^7.4.3"
detect-secrets = "~1.4.0"
django-debug-toolbar = "^3.7.0"
dslr = "^0.4.0"
flake8 = "7.0.0"
honcho = "^1.0.1"
isort = "5.13.2"
mkdocs = "1.3.0"
mkdocs-material = "8.2.8"
pre-commit = "~2.11"
pudb = "^2020.1"
pymdown-extensions = "9.3"
tblib = "^3.0.0"

[build-system]
requires = ["poetry>=1,<2"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.mypy_cache
    | \.tox
    | \.venv
    | __pycache__
    | _build
    | build
    | dist
    | docs
    | venv
    | node_modules
  )/
)
'''
